Module 2/Chapter7/django-myproject-07/requirements.txt
Django==1.8.4
django-cms==3.1.3
djangocms-text-ckeditor
djangocms-file
djangocms-flash
djangocms-googlemap
djangocms-inherit
djangocms-picture
djangocms-snippet
djangocms-teaser
djangocms-video
django-reversion
django-sekizai==0.8.2